refactor: split selfplay utilities
